feat(CSAF2.1): add CVSS check vor cvss4.0
