+++ tests/csaf_2_1/oasis.js
@@ -29,7 +29,6 @@ const excluded = [
   '6.1.55',
   '6.1.56',
   '6.2.11',
-  '6.2.19',
   '6.2.20',
   '6.2.21',
   '6.2.24',
